Make logrotate more intelligent for high rotation rates
Ticket: CM-14604
Reviewed By: mallik,wkok,anuradha,jguy,sharpd,kanna,curt
Testing Done: rotated a number of new and old log files, adjusted dates

The idea here is to catch "new" (less than a day old) log files that are
about to be rotated away by logrotate, using the preremove function,
based on the assumption that it's a case of runaway logging (usually
do to a persistent or repetitive error).

The hope/belief is that in this circumstances, the saved file(s) will
be likely to have the the "first" error messages in them, and we'll
keep them around for up to 14 days, to improve support and dev debugging.

Requires matching change in cumulus-tools for functionality