*: add transitional support for quagga configs
* Add -q flag to all frr binaries that enables quagga compat mode
* Copy frr init script to quagga, put -q where necessary
* Add quagga.service that uses this init script
* Add infobox to read CL docs for upgrade path

Note that vtysh run without -q from the command line runs in normal mode
and not in compatibility mode. In 96.2% of cases this should be fine
since almost no configuration is present in vtysh.conf, and watchfrr &
the init script are both smart enough to pass -q when executing a config
load or write.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>