debian: move debian/ directory to debianpkg/
Upstream did this to appease downstream debian developers (having debian/
in the orig.tar.gz sucks for them). We (cumulus) don't want it downstream,
but I'm applying it for now to get the rest of the patches to apply.

For some reason 3-way merge failed, so I had to remake this manually. Will
correct any possible mistakes down the line.

Signed-off-by: Silas McCroskey <smccroskey@cumulusnetworks.com>