deb-based-distros: clean up leftover debconf references
Testing done: built and installed on all ubuntu

Debconf was in use by this packaging a long time ago, and references to it
were never fully cleaned up when the one question it asked was deleted.

Clears the following lintian warning:
W: frr: no-debconf-templates

cumulus note: Had to reapply this manually. Not sure why git am hates
the .po files so much.

Signed-off-by: Silas McCroskey <smccroskey@cumulusnetworks.com>