lib/command_lex.l: ignore missing function prototypes
The Ubuntu Trusty build fails when attempting to compile the flex-generated
files due to missing prototypes which should be in the flex-generated header
file.

As this issue doesn't show up on Debian Jessie or Ubuntu Xenial and
there's no obvious way to work around it, opted to simply ignore the
warning, following the sign-compare example in the same file.

Signed-off-by: Silas McCroskey <smccroskey@cumulusnetworks.com>