bgpd: Fix VRF route leaking for multipath routes
Ensure that the next hop of the leaked VRF is not overwritten when the
route is being imported into the target VRF from the VPN table. Also, in
the case of multipath routes, ensure that the nexthop's ifindex is not
inadvertently reset.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>
(cherry picked from commit 6488723c062d6e21ae8fd1604d1dd9a4912a65bb)

Ticket: CM-20384
Testing Done:
1. Manual verification
2. Ran vrf_route_leak_test.py - 1 failure to be tracked separately