bgpd: Set NEXT_HOP attribute for EVPN imported routes
Ensure that when EVPN routes are imported into a VRF as IPv4 routes,
the NEXT_HOP attribute is set. In the absence of this, this attribute
is currently not generated when advertising the route to peers in the
VRF. It is to be noted that the source route (the EVPN route) will only
have the MP_REACH_NLRI attribute that contains the next hop in it.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>
Reviewed-by:   Mitesh Kanjariya <mitesh@cumulusnetworks.com>
Reviewed-by:   Don Slice <dslice@cumulusnetworks.com>
(cherry picked from commit 29bea3cb04daffcbfd38fde9b93c748d6c9fdd06)

Ticket: CM-20776
Testing done:
a) Verify my failure scenario with EBGP numbered in core
b) Run Anitha's tests to verify IBGP numbered in core
c) Run evpn-smoke and verify no failures due to this change