Cumulus: Fix FRR 'Conflicts:' back to a versioned 'Breaks:'
Ticket: CM-20841
Testing-done: on 3.3.2 VX:
 removed quagga (cl3u13)
 installed built frr using dpkg
 apt-get installed quagga (cl3u14)
 apt-get upgrade
Reviewed-by: CCR-7464

A merge prior the 3.6.0 release changed our versioned 'Breaks:' of
quagga (which works if you've updated to the version which depends on
quagga-compat) to a non-versioned 'Conflicts:' (which doesn't work
with quagga at all).  This broke upgrades from releases which still
had quagga around (3.3.2 and below).

Fix it back to the way it was before to fix the upgrade issue.