lib: [cumulus] Reorder evpn prefix for IPv6
Move the unused ethernet tag field in the EVPN prefix past the IP
address field so that the relevant portion of the prefix can have
a prefix length less than 255 bytes. This is necessary to ensure
IPv6 hosts on the same VLAN with same MAC and IPv6 address differing
only in their last 2 or 4 bytes are treated as distinct.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>

Ticket: CM-20849, CM-20952
Testing Done: Manual verification