ospfd: fix segfault when originating network-LSA
When FRR is elected as an OSPF DR, in accordance with RFC 2328 ss.
12.4.2 it attempts to originate a network-LSA. The function responsible
for creating a new network-LSA checks that we have at least one full
adjacency, and if we don't, returns NULL (since 12.4.2 says the DR
doesn't originate the network-LSA if there are no full adjacencies). But
the caller doesn't check the returned LSA for nullity before using it
and we get a segfault.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>