zebra: Enhancements to EVPN operational commands
Enhance the EVPN MAC and Neighbor cache display to show additional
information such as the mobility sequence numbers and the state.
Ensure that the neighbor state is set in a couple of places so
that the display is correct.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>
(cherry picked from commit e87d0ba9c4ec8354daceb44899be78e5115d400b)

Conflicts:
	zebra/zebra_vxlan.c