zebra: Allow higher admin distance static routes to be used
Suppose we have two static routes:
ip route 4.3.2.1/32 55.55.55.55
ip route 4.3.2.1/32 192.168.201.44 21

The second one has an admin distance of 21 and should not be installed
unless the first one becomes unreachable.  This was not working
because the static nexthop tracking code in zebra_rnh.c was
intentionally setting the route changed status flag and when
we got into rib_process we saw the route changed status flag
and intentionally were not calling nexthop_active_update()
because static routing already figured it out.

The problem here is that the admin distance was kicking in
and the unreachable route was being selected.

Remove the ROUTE_CHANGED status flag from processing
the static nexthops and just let us do a bit of
non-optimized code in rib_process and let it look
at the nexthops to figure out what is valid or not.

This code is fixed upstream via the staticd change
and as such should not be upstreamed.

Ticket: CM-21379
Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>