zebra: remove `onlink` keyword from bad place
`onlink` may only be specified where both a gateway and an interface are
also specified, per kernel limitations. I accidentally included it on
one of the v4 static route commands that doesn't require a gateway to be
specified, causing some confusion since the `onlink` specifier was
subsequently ignored by the kernel and thus didn't show up in the
config.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>