pimd: prevent ipmr-lo addition to the OIL if SIP is local
In a pim-mlag setup the VTEP can use an individual SIP (27.0.0.11 in the
example) or an anycast SIP (36.0.0.11 in the example). The VTEP switches
between the two SIPs based on the state of the peerlink/ISL. If peerlink
is down individual IP is used and if peerlink is up anycast IP is used.

During the modification the mroute transitions from orig->non-orig->orig.
For orig-mroutes there are additional checks to prevent the addition of the
vxlan termination device (ipmr-lo) to the OIL. This is important as we
don't want to terminate packets that we encapsulate/originate. But the
check is not in effect when the mroute transitions to non-orig. Any
encapsulated (with SIP 27.0.0.11) packets in this window will loop -
root@TORC11:~# ip mr |grep 239.1.1.100|grep "27.0.0.11\|36.0.0.11\|0.0.0.0"
(36.0.0.11,239.1.1.100) Iif: peerlink-3.4094 Oifs: uplink-2 peerlink-3.4094
(0.0.0.0,239.1.1.100)   Iif: uplink-2   Oifs: pimreg ipmr-lo uplink-2
(27.0.0.11,239.1.1.100) Iif: lo         Oifs: pimreg ipmr-lo uplink-2 >>>>
root@TORC11:~#

To avoid those windows additional checks have been added to prevent
addition of ipmr-lo in the OIL of any mroute whose SIP is local. This
guarantees that a locally encapuslated will never be locally decapsulated.

Ticket: CM-25066

Signed-off-by: Anuradha Karuppiah <anuradhak@cumulusnetworks.com>
(cherry picked from commit 7d93ff8f075c8998d5266674e241a70a8eb51c24)