zebra: Have zebra store replay information to pim
PIM when it starts signals to zebra that it should replay
vxlan information to zebra as that it is ready.  Additionally
we store a flag to decide to send vxlan information to PIM.

What was happening is that you could have ordering issues with
PIM starting up before evpn has been turned on in zebra and
we were not sending any data to PIM.

This commit stores the flag irrelevant if evpn is up and running
or not.  When evpn is turned on we will send the data then.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>