staticd: Fix static_nht_update to actually know route we are installing
We are using static_nht_update in two different cases:
1) We have received a callback that a nexthop has changed and we need
to find any static route that is using it and we must refigure it.
2) We have received a new static route that depends on a pre-existing
nexthop, in which case we can go straight to the chase and just refigure
that particular node, since we already have all the information we need.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>