bgpd: Fix route-map check for EVPN advertise-map
The code that applies the EVPN advertise route-map to see if a route
from the ipv4/unicast or ipv6/unicast routing table can be injected
into the l2vpn/evpn routing table had an incorrect check that would
fail as soon as the route-map had any 'set' statements. Fix the
check.

Signed-off-by: Vivek Venkatraman <vivek@cumulusnetworks.com>